reworked function slot manager
