+++ src/main/java/eu/newton/ui/functionmanager/FunctionManager.java
@@ -1,4 +1,4 @@
-package eu.newton.reworkedui.functionmanager;
+package eu.newton.ui.functionmanager;
 
 import eu.newton.MathFunction;
 import eu.newton.api.IDifferentiable;
